split api/client/room
